# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Posts, Alunos e Professores",
    "version": "1.0.0",
    "description": "Documentação da API Express com rotas para Posts, Alunos, Professores e Autenticação."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/posts": {
      "get": {
        "summary": "Listar todos os posts",
        "responses": {
          "200": {
            "description": "Lista de posts"
          }
        }
      },
      "post": {
        "summary": "Criar novo post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post criado"
          }
        }
      }
    },
    "/posts/search": {
      "get": {
        "summary": "Buscar posts por título ou conteúdo",
        "parameters": [
          {
            "name": "titulo",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "conteudo",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts filtrados"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "summary": "Buscar post por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Post encontrado" },
          "404": { "description": "Post não encontrado" }
        }
      },
      "put": {
        "summary": "Atualizar post por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Post" } }
          }
        },
        "responses": {
          "200": { "description": "Post atualizado" }
        }
      },
      "delete": {
        "summary": "Deletar post por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Post deletado" }
        }
      }
    },
    "/alunos": {
      "get": {
        "summary": "Listar alunos",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Lista de alunos" }
        }
      },
      "post": {
        "summary": "Criar novo aluno",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Aluno" } }
          }
        },
        "responses": {
          "201": { "description": "Aluno criado" }
        }
      }
    },
    "/alunos/{id}": {
      "put": {
        "summary": "Atualizar aluno",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Aluno" } }
          }
        },
        "responses": {
          "200": { "description": "Aluno atualizado" }
        }
      },
      "delete": {
        "summary": "Deletar aluno",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Aluno deletado" }
        }
      }
    },
    "/professores": {
      "get": {
        "summary": "Listar professores",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Lista de professores" }
        }
      },
      "post": {
        "summary": "Criar novo professor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Professor" } }
          }
        },
        "responses": {
          "201": { "description": "Professor criado" }
        }
      }
    },
    "/professores/{id}": {
      "put": {
        "summary": "Atualizar professor",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Professor" } }
          }
        },
        "responses": {
          "200": { "description": "Professor atualizado" }
        }
      },
      "delete": {
        "summary": "Deletar professor",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Professor deletado" }
        }
      }
    },
    "/professores/auth": {
      "post": {
        "summary": "Autenticar professor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": { "type": "string" },
                  "senha": { "type": "string" }
                },
                "required": ["login", "senha"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Autenticação bem-sucedida" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Post": {
        "type": "object",
        "properties": {
          "titulo": { "type": "string" },
          "conteudo": { "type": "string" },
          "autor": { "type": "string" }
        },
        "required": ["titulo", "conteudo", "autor"]
      },
      "Aluno": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "curso": { "type": "string" },
          "matricula": { "type": "string" }
        },
        "required": ["nome", "curso", "matricula"]
      },
      "Professor": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "materia": { "type": "string" },
          "login": { "type": "string" },
          "senha": { "type": "string" }
        },
        "required": ["nome", "materia", "login", "senha"]
      }
    }
  }
}